Implemented Advanced Events
New events API which allows an event to be attached to any object.
Documentation is in the Steam Workshop EasyAPI Guide.

I also removed the example code from the top of the script so it is
easier to get started on a new script without having to delete the
example code.  The examples will be available in a subfolder in this
project and in the Steam Workshop Guide.