EasyMenuItem has been refactored to use two classes
There is an issue in the Space Engineers programming block where classes
don't work in Lists.  This caused all kinds of issues with the menu
system since menu items had to be structs which meant that all
properties were copied around constantly.  This change splits the
EasyMenuItem struct into a class and a struct.  All properties are
stored within the class so that everything is now passed by reference.
This makes dynamic sub-menus possible.