Added Modules for script generation
I'm preparing to launch a website which will handle dependencies of
various scripts and combine selected "modules" into one script ready to
be pasted into the PB.  The modules directory now contains the first set
of modules that will be on that site.  People will be able to contribute
modules by logging in and uploading a script or providing a steam
workshop id for their script.  The EasyAPI.cs file at the root of this
repository is considered deprecated.  All future changes should be made
to the modules.